Main

Documentation

main :: IO ()

Given URL as user input, crawls and pretty prints JSON of subdomain structure.

crawlSubdomain

Arguments

:: [Webpage]
The accumulator for visited webpages

-> [Url]
The accumulator for visited URLs

-> [Url]
The list of URLs to still visit

-> IO [Webpage]
Returns final list of visited webpages

Accumulates a list of unique visited webpages and URLs by crawling a given list of URLs. Prints update for each page crawled and for each page not found. Accumulators should usually be initiated as empty lists.

crawlFailure

Arguments

:: [Webpage]
The list of visited webpages

-> [Url]
The list of visited URLs

-> [Url]
The list of this URL and URLs to still visit

-> IO [Webpage]
Returns final list of visited webpages

Outputs a warning message with URL not reached and crawls remaining URLs with this URL marked as seen. There *must be at least one URL to still visit*.

crawlSuccess

Arguments

:: [Webpage]
The list of visited webpages

-> [Url]
The list of visited URLs

-> [Url]
The list of this URL and URLs to still visit

-> ByteString
The source at this URL

-> IO [Webpage]
Returns final list of visited webpages

Outputs message with crawled URL, crawls and saves Webpage, marks URL as seen and crawls remaining URLs, including any URLs marked on this page. There *must be at least one URL to still visit*.