Utils.HtmlParser

Documentation

data Webpage

A container for webpages, made up of the page URL, the URLs of its assets, and tthe URLs of any webpages it links to.

Constructors

Webpage

Fields

url :: Url

The URL for the webpage

assets :: [Url]

The list of URLs of linked assets

links :: [Url]

The list of URLs of linked pages

crawlWebpage

Arguments

:: Url
The URL of the page to be crawled

-> ByteString
The HTML source for the page

-> Webpage
Returns a Webpage object with links

Returns a Webpage object with asset and page links for a given URL and HTML source.

parseHtml

Arguments

:: ByteString
The HTML source to be parsed

-> [Maybe ByteString]
The accumulator for asset links

-> [Maybe ByteString]
The accumulator for page links

-> ([Maybe ByteString], [Maybe ByteString])
Returns tuple containing list of all Maybe asset links and list of all Maybe page links

Given HTML source, accumulates seperately links for assets and other pages. Accumulators should usually be initiated as empty lists. Any links are wrapped in Just, tags without links produce Nothing.

parseAttributes

Arguments

:: ByteString
The bytestring to be parsed

-> (Maybe ByteString, ByteString)
Returns tuple containing Maybe link value and remainder of bytestring to be parsed

Given a bytestring beginning with a list of attributes, returns Just the first value which refers to a link, or if no such value exists, Nothing. Also returns the remainder of the bytestring to be parsed.

parseName

Arguments

:: ByteString
The bytestring to be parsed

-> (ByteString, ByteString)
Returns the name of the first attribute and remainder of bytestring to be parsed

Given a bytestring beginning with an attribute and /with no preceeding whitespace/, returns the name of the attribute and the remainder of the bytestring to be parsed.

parseValue

Arguments

:: ByteString
The bytestring to be parsed

-> (ByteString, ByteString)
Returns the first attribute value and the remainder of the bytestring to be parsed

Given a bytestring beginning with an attribute value and /with no preceeding whitespace/, returns the attribute value and the remainder of the bytestring to be parsed.

dropSpace :: ByteString -> ByteString

Removes any whitespace from the beginning of a bytestring

dropEquals :: ByteString -> ByteString

Removes any equal signs from the beginning of a bytestring

dropQuote :: ByteString -> ByteString

Removes any quotation marks from the beginning of a bytestring

assetTags :: [ByteString]

HTML tags that contain links to assets

linkTags :: [ByteString]

HTML tags that contain links to pages

attributeNames :: [ByteString]

HTML attribute names that refer to assets and links