Utils.UrlParser

Documentation

data Url

A container for URLs, made up of scheme, authority and path.

Constructors

Url

Fields

scheme :: ByteString

E.g. "http:"

authority :: ByteString

E.g. "www.example.com"

path :: ByteString

E.g. "dirpage.html"

formatMaybeAssetUrls

Arguments

:: [Maybe ByteString]
List of Maybe asset URLs to parse

-> Url
The current URL

-> [Url]
Returns list of parsed asset URLs

Parses a list of Maybe linked URLs relative to a given current URL. Returns a list of URL objects for valid asset links.

formatMaybeLinkedUrls

Arguments

:: [Maybe ByteString]
List of Maybe linked URLs to parse

-> Url
The current URL

-> [Url]
Returns list of parsed linked URLs

Parses a list of Maybe linked URLs relative to a given current URL. Returns a list of URL objects for valid links within the same subdomain.

parseUrl

Arguments

:: ByteString
The bytestring URL to be parsed

-> Url
Returns the parsed Url

Parses a bytestring representation of a URL into a Url object. Scheme is "http:" unless specified as "https:". Returns "http:////" for any other scheme. Will always produce an absolute URL, including authority. Returned scheme and authority are lowercase and path is normalised.

parseUrlRelativeTo

Arguments

:: ByteString
The linked URL, to be parsed

-> Url
The current URL (parent)

-> Url
Returns the linked URL parsed relative to the current URL

Parses a bytestring for a linked URL into a Url object, by considering it relative to a given current URL. If no scheme or authority is specified for the linked URL, uses the current URL to guess. The linked URL will be considered relative (and the path added to that of the current url) iff it: - does not begin with "www." - AND does not have a / except for as the first character or does not have a . before a / except for as the first character Returned scheme and authority are lowercase and path is normalised.

parseScheme

Arguments

:: ByteString
The URL to be parsed

-> (Maybe ByteString, ByteString)
Returns tuple of Maybe scheme and remainder of URL

Parses a bytestring to retrieve its scheme if it is "http:" or "https:" and the remainder to be parsed. If no scheme is found, returns Nothing and the entire URL. If an incorrect scheme is found (e.g. "mailto:") returns Nothing and an empty string.

parseAuthority

Arguments

:: ByteString
The URL (without scheme) to be parsed

-> (Maybe ByteString, ByteString)
Returns tuple of Maybe authority and remainder of the URL

Parses a bytestring with scheme removed, to retrieve its authority and the and. If no authority is found, returns Nothing and the entire URL.

parseAuthority' :: ByteString -> (Maybe ByteString, ByteString)

Helper function for parseAuthority which always returns Url with authority.

parsePath

Arguments

:: ByteString
The URL (without scheme and authority) to be be parsed

-> ByteString
Returns path without queries and fragments

Parses a bytestring with scheme and authority removed, to retrieve its path, removing queries and fragments.

normalise

Arguments

:: ByteString
Absolute path to be normalised, E.g. "dir../index.html"

-> ByteString
Returns normalised path, E.g. "/"

Normalises a path by removing dot-segments and references to index.html Path *must be absolute* (begins with /).

normalise' :: ByteString -> [ByteString] -> [ByteString]

Helper function for normalise

endAuthChars :: String

Characters which end of authority

endPathChars :: String

Characters which signify end of path

isValidLinkedUrl

Arguments

:: Url
The linked URL

-> Url
The current URL to compare agains

-> Bool
Returns whether URLs have the same authority

Returns true if two URLs have the same subdomain (authority).

showUrl

Arguments

:: Url
Url object to be shown

-> ByteString
Returns URL as lazy bytestring

Shows a Url object as a lazy bytestring.