Main

Documentation

main :: IO ()

Given URL as user input, crawls and pretty prints JSON of subdomain structure. Input can be given with or without scheme. If no scheme is provided, http will be selected by default. Output JSON is a list of viewable webpages, each with its URL and a list of any assets. Example:

>>> main
Please provide a starting URL:
>>> www.haskell.org
[
{
    "url": "http://www.haskell.org/",
    "assets": [
        "http://www.haskell.org/static/js/tryhaskell.pages.js",
        "http://www.haskell.org/static/js/tryhaskell.js",
        "http://www.haskell.org/static/js/jquery.console.js",
        "http://www.haskell.org/static/js/home.js",
        "http://www.haskell.org/static/js/bootstrap.min.js",
        "http://www.haskell.org/static/js/jquery.js",
        "http://www.haskell.org/static/img/rackspace.svg",
        "https://i.vimeocdn.com/video/452269027_150x84.jpg",
        "https://i.vimeocdn.com/video/456929840_150x84.jpg",
        "https://i.vimeocdn.com/video/456929997_150x84.jpg",
        "https://i.vimeocdn.com/video/469227196_150x84.jpg",
        "https://i.vimeocdn.com/video/469235326_150x84.jpg",
        "https://i.vimeocdn.com/video/476988542_150x84.jpg",
        "http://www.haskell.org/static/img/haskell-logo.svg",
        "http://www.haskell.org/static/css/hl.min.css",
        "https://fonts.googleapis.com/css",
        "http://www.haskell.org/static/img/favicon.ico"
    ]
},
... (all reachable pages)
]

crawlSubdomain

Arguments

:: Url
The URL to start crawling from

-> IORef [Url]
A list of seen URLs to be excluded

-> IO [Webpage]
Returns a list of all reachable Webpages

Crawls the subdomain of the given URL, returning a list of Webpages of all the reachable pages from the given URL, excluding any which are at URLs in the given list of seen URLs.

crawlFailure

Arguments

:: Url
The URL that could not be reached.

-> IO [Webpage]
Returns an empty list.

Outputs a warning message with the URL that could not be reached. Returns an empty list of Webpages.

crawlSuccess

Arguments

:: ByteString
The HTML source of the given URL

-> Url
The URL being crawled

-> IORef [Url]
A list of seen URLs to be excluded

-> IO [Webpage]
Returns a list of all reachable Webpages

Outputs a message with the URL being crawled. Crawls the webpage, updates the seen URLs with any unseen URLs present on that page, and concurrently crawls the unseen URLs. Returns all reachable pages that haven't been seen yet.

updateUrls

Arguments

:: [Url]
A list of crawled URLs to be processed

-> [Url]
The list of URLs seen so far

-> ([Url], [Url])
Returns a tuple containing the new list of seen URLs, and a list of unseen URLs

Given a list of URLs, produces a tuple conataining an updated list of seen URLs and a list of any unseen URLs.